refactor: rename main file
