--- src/index.ts
@@ -1,3 +1,0 @@
-import { startServer } from './http/server';
-
-startServer();
