+++ src/main.ts
@@ -1,0 +1,3 @@
+import { startServer } from './http/server/server';
+
+startServer();
